Initial implementation of String Literal Equals String (SLEQS).
This change includes the main function (sleqs) implementation along with  few helper routines located in sleqs.d file. In this file there is also  a memcmp based facility with the same interface as sleqs mostly used for comparison of the generated code but also potentially for consumers of the library. Finally there are a couple of basic performance tests which measure the performance of the sleqs and memcmp functions against 127 compile time generated strings with sizes varying from 1 to 128 bytes.